fileOne/requirements.txt
tensorflow
Flask
Pillow
requests